refactor: moved tests to job factory
